# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Briefed API",
    "description": "Personal AI email agent. Runs a daily pipeline on the user's Gmail inbox, categorizes and prioritizes messages, and summarizes tech news via the Claude API.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/rubric": {
      "get": {
        "tags": [
          "rubric"
        ],
        "summary": "List classification rules",
        "description": "Return every rubric rule the authenticated user owns.\n\nArgs:\n    user_id: Authenticated owner.\n    session: Active async session.\n\nReturns:\n    :class:`RubricRulesListResponse` with rules ordered by\n    ``priority DESC``.",
        "operationId": "list_rules_api_v1_rubric_get",
        "parameters": [
          {
            "name": "briefed_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Briefed Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricRulesListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rubric"
        ],
        "summary": "Create a classification rule",
        "description": "Insert a new rule for the authenticated user.\n\nArgs:\n    payload: Request body.\n    user_id: Authenticated owner.\n    session: Active async session.\n\nReturns:\n    The created :class:`RubricRuleOut`.",
        "operationId": "create_rule_api_v1_rubric_post",
        "parameters": [
          {
            "name": "briefed_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Briefed Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RubricRuleIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricRuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rubric/{rule_id}": {
      "put": {
        "tags": [
          "rubric"
        ],
        "summary": "Replace a classification rule",
        "description": "Replace ``rule_id``'s predicate + action, bumping ``version``.\n\nArgs:\n    rule_id: Target rule.\n    payload: Request body.\n    user_id: Authenticated owner.\n    session: Active async session.\n\nReturns:\n    The updated :class:`RubricRuleOut`.\n\nRaises:\n    HTTPException: 404 when the rule does not belong to the caller.",
        "operationId": "update_rule_api_v1_rubric__rule_id__put",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          },
          {
            "name": "briefed_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Briefed Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RubricRuleIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricRuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "rubric"
        ],
        "summary": "Delete a classification rule",
        "description": "Hard-delete ``rule_id``.\n\nArgs:\n    rule_id: Target rule.\n    user_id: Authenticated owner.\n    session: Active async session.\n\nRaises:\n    HTTPException: 404 when the rule does not belong to the caller.",
        "operationId": "delete_rule_api_v1_rubric__rule_id__delete",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          },
          {
            "name": "briefed_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Briefed Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RubricRuleIn": {
        "properties": {
          "priority": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Priority",
            "default": 100
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "default": "Untitled rule"
          },
          "match": {
            "additionalProperties": true,
            "type": "object",
            "minProperties": 1,
            "title": "Match"
          },
          "action": {
            "additionalProperties": true,
            "type": "object",
            "title": "Action"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "match",
          "action"
        ],
        "title": "RubricRuleIn",
        "description": "Request body for POST / PUT.\n\nAttributes:\n    priority: Higher wins. Defaults to ``100``.\n    name: Human-readable settings label.\n    match: Predicate dict; keys limited to\n        :data:`_ALLOWED_MATCH_KEYS`.\n    action: Verdict dict; must contain ``label`` \u2208\n        :data:`_ALLOWED_LABELS` and ``confidence`` \u2208 ``[0, 1]``.\n    active: Soft-delete switch. Defaults to ``True``."
      },
      "RubricRuleOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "match": {
            "additionalProperties": true,
            "type": "object",
            "title": "Match"
          },
          "action": {
            "additionalProperties": true,
            "type": "object",
            "title": "Action"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "priority",
          "match",
          "action",
          "version",
          "active",
          "created_at",
          "updated_at"
        ],
        "title": "RubricRuleOut",
        "description": "Response body for list / create / update.\n\nAttributes:\n    id: Rule primary key.\n    name: Human-readable settings label.\n    priority: Priority plumbed through from the DB row.\n    match: Predicate dict.\n    action: Verdict dict.\n    version: Monotonically increasing version.\n    active: Soft-delete switch.\n    created_at: When the rule was first inserted.\n    updated_at: Last-modified timestamp."
      },
      "RubricRulesListResponse": {
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RubricRuleOut"
            },
            "type": "array",
            "title": "Rules",
            "default": []
          }
        },
        "type": "object",
        "title": "RubricRulesListResponse",
        "description": "Envelope for ``GET /rubric``.\n\nAttributes:\n    rules: Every :class:`RubricRuleOut` owned by the caller,\n        ordered by ``priority DESC``."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}